image rotation modify

+++ Test/Pistol.cpp
@@ -6,7 +6,7 @@ Pistol::Pistol() :Weapon()
 
 }
 
-void Pistol::draw_weapon(HDC mDC, const Vector2D<int>& center)
+void Pistol::draw_weapon(HDC mDC, const Vector2D<float>& center)
 {
 
 }
