+++ Test/Pistol.h
@@ -6,7 +6,7 @@ class Pistol
 {
 public:
 	Pistol();
-	virtual void draw_weapon(HDC mDC, const Vector2D<int>& center);
+	virtual void draw_weapon(HDC mDC, const Vector2D<float>& center);
 	virtual void update();
 	virtual void attack();
 };
